% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cell07PNs}
\alias{Cell07PNs}
\title{Standard neuronlist object containing neurons from nat::Cell07PNs}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 40.}
\usage{
Cell07PNs
}
\description{
Standard neuronlist object containing neurons from nat::Cell07PNs
}
\details{
This is just a straight copy of the \code{\link[nat]{Cell07PNs}}
  object.
}
\keyword{datasets}
