% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pnsfromnlfh}
\alias{pnsfromnlfh}
\title{Neuron list constructed from neuronlistfh with delayedAssign}
\description{
Neuron list constructed from neuronlistfh with delayedAssign
}
\details{
The source data for this object live in \code{extdata} and consist
  of a metadata object saved as \code{{pnsnlfh.rds}} along with a \code{data}
  folder. When the package is loaded, an in memory version of the data is
  constructed using \code{\link{delayedAssign}}.
}
