% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pnsnlfh}
\alias{pnsnlfh}
\title{neuronlistfh object made with delayedAssign}
\description{
neuronlistfh object made with delayedAssign
}
\details{
The source data for this object live in \code{extdata} and consist
  of a metadata object saved as \code{{pnsnlfh.rds}} along with a \code{data}
  folder. When the package is loaded the neuronlistfh object is loaded into
  an object called \code{pnsnlfh} using \code{\link{delayedAssign}}.
}
